% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{Operend coercion methods}
\alias{Operend coercion methods}
\title{Coerce Operend object classes}
\description{
These methods coerce Operend S4 objects to objects of other classes, and
vice versa.
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
