% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Entities.R
\name{Entities}
\alias{Entities}
\alias{addEntity}
\alias{deleteEntity}
\alias{getEntity}
\alias{listEntities}
\alias{updateEntity}
\title{Add, delete, retrieve, update, or list Entity record(s)}
\usage{
addEntity(
  class,
  id,
  variables,
  permissions,
  verbosity = getOption("opeRend")$verbosity
)

deleteEntity(id, verbosity = getOption("opeRend")$verbosity)

getEntity(id, verbosity = getOption("opeRend")$verbosity)

listEntities(class, variables = list())

updateEntity(
  id,
  variables,
  permissions,
  verbosity = getOption("opeRend")$verbosity
)
}
\arguments{
\item{class}{A character string specifying the class of the \code{Entity} record}

\item{id}{A character string specifying the unique identifier of an \code{Entity}
record to be added, deleted, retrieved, or updated}

\item{variables}{Additional arguments specifying variables of the \code{Entity}
record to be added or updated, or variables on which to limit a listing}

\item{permissions}{An optional \code{\linkS4class{operendPermissions}} object
specifying the permissions to be used when creating or updating the record}

\item{verbosity}{An integer-coercible value specifying the verbosity level:
\describe{
  \item{\code{0}}{Do not print messages}
  \item{\code{1}}{Print only high-level messages}
  \item{\code{2}}{Show headers}
  \item{\code{3}}{Show headers and bodies}
  \item{\code{4+}}{Show headers, bodies, and curl status messages}
}
Defaults to \code{getOption("opeRend")$verbosity}.}
}
\value{
\describe{
  \item{\code{addEntity}, \code{getEntity}, \code{updateEntity}}{
    If the operation is successful, an \code{\linkS4class{operendEntity}}
    object.
  }
  \item{\code{deleteEntity}}{
    A logical value stating whether the operation was successful (invisibly).
  }
  \item{\code{listEntities}}{
    An \code{\linkS4class{operendEntityList}} object.
  }
  \item{All functions}{
    If an error is encountered, the function terminates with a message.
  }
}
}
\description{
These functions attempt to add, delete, retrieve or update an \code{Entity}
record, or to list \code{Entity} records.
}
\details{
If \code{updateEntity} is called without argument \code{variables},
the argument \code{permissions} \emph{must} be provided.
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
