% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EntityClasses.R
\name{EntityClasses}
\alias{EntityClasses}
\alias{addEntityClass}
\alias{deleteEntityClass}
\alias{getEntityClass}
\alias{listEntityClasses}
\alias{updateEntityClass}
\title{Add, delete, retrieve, update, or list EntityClass record(s)}
\usage{
addEntityClass(file, permissions, verbosity = getOption("opeRend")$verbosity)

deleteEntityClass(name, verbosity = getOption("opeRend")$verbosity)

getEntityClass(name, verbosity = getOption("opeRend")$verbosity)

listEntityClasses()

updateEntityClass(
  name,
  file,
  permissions,
  verbosity = getOption("opeRend")$verbosity
)
}
\arguments{
\item{file}{One of the following:
\itemize{
  \item{
    a character string containing the full path to a file
    containing an \code{EntityClass} definition in JSON
  }
  \item{
    a \code{\link{connection}} (which will be opened if necessary, and closed
    at the end of the function call) to a file
    containing an \code{EntityClass} definition in JSON
  }
}}

\item{permissions}{An optional \code{\linkS4class{operendPermissions}} object
specifying the permissions to be used when creating or updating the record}

\item{verbosity}{An integer-coercible value specifying the verbosity level:
\describe{
  \item{\code{0}}{Do not print messages}
  \item{\code{1}}{Print only high-level messages}
  \item{\code{2}}{Show headers}
  \item{\code{3}}{Show headers and bodies}
  \item{\code{4+}}{Show headers, bodies, and curl status messages}
}
Defaults to \code{getOption("opeRend")$verbosity}.}

\item{name}{A character string specifying the name of an \code{EntityClass}
record to be deleted, retrieved, or updated}
}
\value{
\describe{
  \item{
    \code{addEntityClass}, \code{getEntityClass}, \code{updateEntityClass}
  }{
    If the operation is successful,
    an \code{\linkS4class{operendEntityClass}} object.
  }
  \item{\code{deleteEntityClass}}{
    A logical value stating whether the operation was successful (invisibly).
  }
  \item{\code{listEntityClasses}}{
    An \code{\linkS4class{operendEntityClassList}} object.
  }
  \item{All functions}{
    If an error is encountered, the function terminates with a message.
  }
}
}
\description{
These functions attempt to add, delete, retrieve or update an
\code{EntityClass} record, or to list \code{EntityClass} records.
}
\details{
If \code{updateEntityClass} is called without argument \code{file}, both of
the arguments \code{name} and \code{permissions} \emph{must} be provided.
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
