% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JobRuns.R
\name{JobRuns}
\alias{JobRuns}
\alias{deleteJobRun}
\alias{getJobRun}
\alias{listJobRuns}
\title{Delete, retrieve, or list JobRun record(s)}
\usage{
deleteJobRun(id, verbosity = getOption("opeRend")$verbosity)

getJobRun(id)

listJobRuns(jobType = "")
}
\arguments{
\item{id}{An integer-coercible value specifying the unique identifier of a
\code{JobRun} record.}

\item{verbosity}{A value coercible to a nonnegative integer, specifying the verbosity level.
A value of FALSE, TRUE, 0 or 1 does not produce any messages.
A value of 2 instructs the curl calls to produce verbose output.
A value greater than 2 produces additional output.
Defaults to \code{getOption("opeRend")$verbosity}.}

\item{jobType}{A character string specifying the type of JobRuns to list.}
}
\value{
\describe{
  \item{\code{getJobRun}}{
    If the operation is successful,
    an \code{\linkS4class{operendJobRun}} object.
  }
  \item{\code{deleteJobRun}}{
    A logical value stating whether the operation was successful (invisibly).
  }
  \item{\code{listJobRuns}}{
    An \code{\linkS4class{operendJobRunList}} object.
  }
  \item{All functions}{
    If an error is encountered, the function terminates with a message.
  }
}
}
\description{
These functions attempt to delete or retrieve a \code{JobRun} record
or to list \code{JobRun} records.
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
