% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{objectId}
\alias{objectId}
\alias{objectId,operendEntity-method}
\alias{objectId,operendEntityClass-method}
\alias{objectId,operendGroup-method}
\alias{objectId,operendJobRun-method}
\alias{objectId,operendToken-method}
\alias{objectId,operendUser-method}
\alias{objectId,operendWorkFileProperties-method}
\title{Retrieve the identifier of an Operend object}
\usage{
objectId(object)

\S4method{objectId}{operendEntity}(object)

\S4method{objectId}{operendEntityClass}(object)

\S4method{objectId}{operendGroup}(object)

\S4method{objectId}{operendJobRun}(object)

\S4method{objectId}{operendToken}(object)

\S4method{objectId}{operendUser}(object)

\S4method{objectId}{operendWorkFileProperties}(object)
}
\arguments{
\item{object}{An Operend R S4 object}
}
\value{
The unique identifier is returned, of the following class:
\describe{
  \item{\code{\linkS4class{operendEntity}}}{
    \code{character}
  }
  \item{\code{\linkS4class{operendEntityClass}}}{
    \code{character}
  }
  \item{\code{\linkS4class{operendGroup}}}{
    \code{character}
  }
  \item{\code{\linkS4class{operendJobRun}}}{
    \code{integer}
  }
  \item{\code{\linkS4class{operendToken}}}{
    \code{character}
  }
  \item{\code{\linkS4class{operendUser}}}{
    \code{character}
  }
  \item{\code{\linkS4class{operendWorkFileProperties}}}{
    \code{integer}
  }
}
}
\description{
This is a generic function for retrieving the primary identifier associated
with an Operend S4 object.
}
\section{Methods}{

Class-specific methods exist for:
\itemize{
  \item{\code{\linkS4class{operendEntity}}}
  \item{\code{\linkS4class{operendEntityClass}}}
  \item{\code{\linkS4class{operendGroup}}}
  \item{\code{\linkS4class{operendJobRun}}}
  \item{\code{\linkS4class{operendUser}}}
  \item{\code{\linkS4class{operendWorkFileProperties}}}
}
}

\author{
Adam C. Gower \email{agower@bu.edu}
}
