% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{objectPermissions}
\alias{objectPermissions}
\alias{objectPermissions,operendEntity-method}
\alias{objectPermissions,operendEntityClass-method}
\alias{objectPermissions,operendJobRun-method}
\alias{objectPermissions,operendWorkFileProperties-method}
\title{Retrieve the permissions of an Operend object}
\usage{
objectPermissions(object)

\S4method{objectPermissions}{operendEntity}(object)

\S4method{objectPermissions}{operendEntityClass}(object)

\S4method{objectPermissions}{operendJobRun}(object)

\S4method{objectPermissions}{operendWorkFileProperties}(object)
}
\arguments{
\item{object}{An Operend R S4 object}
}
\value{
An \code{\linkS4class{operendPermissions}} object.
}
\description{
This is a generic function for retrieving the permissions associated
with an Operend S4 object.
}
\section{Methods}{

Class-specific methods exist for:
\itemize{
  \item{\code{\linkS4class{operendEntity}}}
  \item{\code{\linkS4class{operendEntityClass}}}
  \item{\code{\linkS4class{operendJobRun}}}
  \item{\code{\linkS4class{operendWorkFileProperties}}}
}
}

\author{
Adam C. Gower \email{agower@bu.edu}
}
