% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opeRend-package.R
\docType{package}
\name{opeRend-package}
\alias{opeRend}
\alias{opeRend-package}
\title{opeRend: R interface to an Operend Core instance}
\description{
Provides R functions for accessing and manipulating data in an Operend Core instance.
}
\author{
\strong{Maintainer}: Adam Gower \email{agower@bu.edu}

}
\keyword{internal}
