% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operendApiCall.R
\name{operendApiCall}
\alias{operendApiCall}
\title{Perform an Operend API call}
\usage{
operendApiCall(
  path,
  query,
  baseUrl = getOption("opeRend")$api_base_url,
  method = c("GET", "POST", "PUT", "DELETE"),
  accept = c("application/json", "application/octet-stream"),
  content,
  contentType = c("application/json", "application/octet-stream"),
  contentIsFilePath = FALSE,
  simplifyDataFrame = FALSE,
  verbosity = getOption("opeRend")$verbosity
)
}
\arguments{
\item{path}{A character vector denoting values to be used to construct the path}

\item{query}{An optional named list with one item (value) for each query parameter, with
the names of the list items corresponding to the names of the parameters}

\item{baseUrl}{A character string specifying the API base URL;
defaults to \code{getOption("opeRend")$api_base_url}}

\item{method}{A character string specifying the HTTP method to use to submit the API call}

\item{accept}{A character string specifying the "Accept" HTTP header to be passed to
\code{req_perform}; defaults to \code{"application/json"}}

\item{content}{An optional variable specifying the content to be uploaded during a POST or
PUT request}

\item{contentType}{A character string specifying the "Content-Type" HTTP header to be passed to
\code{req_perform}; defaults to \code{"application/json"}}

\item{contentIsFilePath}{A logical variable specifying whether \code{content} contains a file path}

\item{simplifyDataFrame}{A logical variable to be passed to \code{\link[jsonlite]{fromJSON}}
specifying whether to coerce JSON arrays containing only records
(JSON objects) into a data frame; defaults to FALSE}

\item{verbosity}{An integer-coercible value specifying the verbosity level:
\describe{
  \item{\code{0}}{Do not print messages}
  \item{\code{1}}{Print only high-level messages}
  \item{\code{2}}{Show headers}
  \item{\code{3}}{Show headers and bodies}
  \item{\code{4+}}{Show headers, bodies, and curl status messages}
}
Defaults to \code{getOption("opeRend")$verbosity}.}
}
\value{
If an error is encountered, an error is thrown,
containing a message constructed from the HTTP response body.
Otherwise:
\itemize{
  \item{
    If the \code{Content-Type} of the HTTP response is set to
    \code{'application/json'}, the response body is converted from JSON to an
    R object, and that object is returned.
  }
  \item{
    If the \code{Content-Type} is set to anything else (or is NA),
    the HTTP response body is returned as is.
  }
}
}
\description{
This function first constructs an Operend API URL with specified parameters
and then uses \code{\link[httr2]{req_perform}} to carry out the call,
converting all input and output to/from JSON as needed.
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
