% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operendApiUrl.R
\name{operendApiUrl}
\alias{operendApiUrl}
\title{Construct an API URL}
\usage{
operendApiUrl(
  ...,
  apiBaseUrl = getOption("opeRend")$api_base_url,
  query = list()
)
}
\arguments{
\item{\dots}{Atomic values denoting items to be used to construct the path}

\item{apiBaseUrl}{A character string specifying the API base URL;
defaults to \code{getOption("opeRend")$api_base_url}}

\item{query}{An optional named list with one item (value) for each query parameter, with
the names of the list items corresponding to the names of the parameters}
}
\value{
A character string containing the specified API URL.
}
\description{
This function constructs a URL from the API base URL and any arguments
(e.g., function names).
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
