% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operendConfig.R
\name{getOperendConfig}
\alias{getOperendConfig}
\title{Operend configuration}
\usage{
getOperendConfig(
  configFile = Sys.getenv("OPEREND_CONFIG_FILE"),
  config = Sys.getenv("OPEREND_CONFIG")
)
}
\arguments{
\item{configFile}{A character string specifying the full path to an INI-formatted config file.
Defaults to the contents of the environment variable
\code{OPEREND_CONFIG_FILE}; if that is not set, default parameters are used.}

\item{config}{A character string specifying the section of the config file to be used.
Defaults to the contents of the environment variable
\code{OPEREND_CONFIG}; if that is not set, \code{"default"}}
}
\value{
A list of configuration parameters, invisibly.
}
\description{
This function parses an Operend config file and extracts a list of
parameters.
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
