% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{operendEntity-class}
\alias{operendEntity-class}
\alias{operendEntity}
\title{Class to contain Entity records}
\description{
This class is the S4 representation of the Entity record type.
It is a container for structured metadata and references to other Operend
records.
}
\section{Slots}{

\describe{
\item{\code{_class}}{A character string specifying the Entity class of the record.
Defaults to \code{""}.}

\item{\code{_entity_id}}{A character string specifying the unique identifier of the record.}

\item{\code{_creation_date}}{A character string specifying the time and date at which the record was
created. Defaults to \code{""}.}

\item{\code{_jobruns}}{An integer vector containing the IDs of any JobRuns associated with the
record; used internally by the server-side app.
Defaults to an empty vector.}

\item{\code{_updated}}{A character string specifying the time and date at which the record was last
updated. Defaults to \code{""}.}

\item{\code{_workfiles}}{An integer vector containing the IDs of any WorkFiles associated with the
record; used internally by the server-side app.
Defaults to an empty vector.}

\item{\code{_creator}}{A character string specifying the \code{username} of the User that
created the record. Defaults to \code{""}.}

\item{\code{_owner}}{A character string specifying the \code{username} of the User that
owns the record. Defaults to \code{""}.}

\item{\code{_permissions}}{An \code{\linkS4class{operendPermissions}} object specifying the permissions
associated with the record. Defaults to read-only.}

\item{\code{.Data}}{A list of objects corresponding to the variables of the Entity.}
}}

\section{Extends}{

This class directly extends class \code{\linkS4class{SimpleList}}.
}

\author{
Adam C. Gower \email{agower@bu.edu}
}
