% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{operendEntityClassList-class}
\alias{operendEntityClassList-class}
\alias{operendEntityClassList}
\title{Class to contain a list of EntityClass objects}
\usage{
operendEntityClassList(...)
}
\arguments{
\item{\dots}{Objects of class \code{\linkS4class{operendEntityClass}}}
}
\description{
This class is a container for one or more EntityClass objects.
}
\section{Extends}{

This class directly extends class \code{\linkS4class{SimpleList}}.
}

\author{
Adam C. Gower \email{agower@bu.edu}
}
