% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{operendEntityList-class}
\alias{operendEntityList-class}
\alias{operendEntityList}
\title{Class to contain a list of Entity objects}
\usage{
operendEntityList(...)
}
\arguments{
\item{\dots}{Objects of or extending class \code{\linkS4class{operendEntity}}}
}
\description{
This class is a container for one or more Entity objects.
}
\section{Extends}{

This class directly extends class \code{\linkS4class{SimpleList}}.
}

\author{
Adam C. Gower \email{agower@bu.edu}
}
