% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{operendGroup-class}
\alias{operendGroup-class}
\alias{operendGroup}
\title{Class to contain data about a Group}
\description{
This class is the S4 representation of the Group record type.
It is a container for data about a given group of Users.
}
\section{Slots}{

\describe{
\item{\code{name}}{A character string specifying the name of the group.
Defaults to \code{""}.}

\item{\code{users}}{An \code{\linkS4class{operendUserList}} specifying the users in the group.
Defaults to an empty \code{operendUserList}.}
}}

\seealso{
Functions for working with Group records are described in
\code{\link{Groups}}.
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
