% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{operendGroupList-class}
\alias{operendGroupList-class}
\alias{operendGroupList}
\title{Class to contain a list of Group objects}
\usage{
operendGroupList(...)
}
\arguments{
\item{\dots}{Objects of or extending class \code{\linkS4class{operendGroup}}}
}
\description{
This class is a container for one or more Group objects.
}
\section{Extends}{

This class directly extends class \code{\linkS4class{SimpleList}}.
}

\author{
Adam C. Gower \email{agower@bu.edu}
}
