% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{operendJobRun-class}
\alias{operendJobRun-class}
\alias{operendJobRun}
\title{Class to contain data about a JobRun}
\description{
This class is the S4 representation of the JobRun record type.
It is a container for data about a given job run.
}
\section{Slots}{

\describe{
\item{\code{creator}}{A character string specifying the creator of the JobRun}

\item{\code{parameters}}{An \code{\linkS4class{operendJobRunParameters}} object}

\item{\code{inputWorkFileIds}}{An \code{\linkS4class{operendWorkFileIdList}} object}

\item{\code{outputWorkFileIds}}{An \code{\linkS4class{operendWorkFileIdList}} object}

\item{\code{allGeneratedWorkFileIds}}{An integer vector}

\item{\code{statusHistory}}{An \code{\linkS4class{operendJobRunStatusHistory}} object}

\item{\code{messageLog}}{A \code{operendJobRunMessageLog} object}

\item{\code{status}}{A character string specifying the status of the JobRun.
Defaults to \code{""}.}

\item{\code{id}}{An integer value}

\item{\code{jobTypeName}}{A character string specifying the type of the JobRun.
Defaults to \code{""}.}

\item{\code{subtasks}}{An \code{\linkS4class{operendJobRunSubtaskList}} object}

\item{\code{permissions}}{An \code{\linkS4class{operendPermissions}} object specifying the permissions
associated with the record. Defaults to read-only.}
}}

\author{
Adam C. Gower \email{agower@bu.edu}
}
