% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{operendJobRunList-class}
\alias{operendJobRunList-class}
\alias{operendJobRunList}
\title{Class to contain a list of JobRun objects}
\usage{
operendJobRunList(...)
}
\arguments{
\item{\dots}{Objects of class \code{\linkS4class{operendJobRun}}}
}
\description{
This class is a container for one or more JobRun objects.
}
\section{Extends}{

This class directly extends class \code{\linkS4class{SimpleList}}.
}

\author{
Adam C. Gower \email{agower@bu.edu}
}
