% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{operendJobRunMessage-class}
\alias{operendJobRunMessage-class}
\alias{operendJobRunMessage}
\title{Class to contain a JobRun message}
\description{
This class is a container for a message associated with a JobRun.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{A character vector of length 2 specifying the message and its timestamp,
respectively.}
}}

\author{
Adam C. Gower \email{agower@bu.edu}
}
