% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{operendJobRunMessageLog-class}
\alias{operendJobRunMessageLog-class}
\alias{operendJobRunMessageLog}
\title{Class to contain a JobRun message log}
\usage{
operendJobRunMessageLog(...)
}
\arguments{
\item{\dots}{Objects of class \code{\linkS4class{operendJobRunMessage}}}
}
\description{
This class is a container for one or more
\code{\linkS4class{operendJobRunMessage}} objects.
}
\section{Extends}{

This class directly extends class \code{\linkS4class{SimpleList}}.
}

\author{
Adam C. Gower \email{agower@bu.edu}
}
