% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{operendJobRunParameters-class}
\alias{operendJobRunParameters-class}
\alias{operendJobRunParameters}
\title{Class to contain parameters for a JobRun}
\usage{
operendJobRunParameters(...)
}
\arguments{
\item{\dots}{Objects of class \code{character}}
}
\description{
This S4 class is a container for a \code{\linkS4class{SimpleList}} of
character vectors containing parameters associated with a JobRun.
Defaults to an empty \code{SimpleList}.
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
