% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{operendJobRunStatusHistory-class}
\alias{operendJobRunStatusHistory-class}
\alias{operendJobRunStatusHistory}
\title{Class to contain a JobRun status history}
\usage{
operendJobRunStatusHistory(...)
}
\arguments{
\item{\dots}{Objects of class \code{\linkS4class{operendJobRunStatusHistoryPoint}}}
}
\description{
This class is a container for one or more
\code{\linkS4class{operendJobRunStatusHistoryPoint}} objects.
}
\section{Extends}{

This class directly extends class \code{\linkS4class{SimpleList}}.
}

\author{
Adam C. Gower \email{agower@bu.edu}
}
