% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{operendJobRunStatusHistoryPoint-class}
\alias{operendJobRunStatusHistoryPoint-class}
\alias{operendJobRunStatusHistoryPoint}
\title{Class to contain a JobRun status history point}
\description{
This class is a container for a status history point
associated with a JobRun.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{A character vector of length 2 specifying the status and its timestamp,
respectively.}
}}

\author{
Adam C. Gower \email{agower@bu.edu}
}
