% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{operendJobRunSubtask-class}
\alias{operendJobRunSubtask-class}
\alias{operendJobRunSubtask}
\title{Class to contain a JobRun subtask}
\description{
This class is a container for a subtask of a JobRun.
}
\section{Slots}{

\describe{
\item{\code{inputWorkFileIds}}{An \code{\linkS4class{operendWorkFileIdList}} object}

\item{\code{outputWorkFileIds}}{An \code{\linkS4class{operendWorkFileIdList}} object}

\item{\code{name}}{A character string specifying the name of the JobRun}

\item{\code{comments}}{A character string containing any comments about the JobRun}
}}

\author{
Adam C. Gower \email{agower@bu.edu}
}
