% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{operendJobRunSubtaskList-class}
\alias{operendJobRunSubtaskList-class}
\alias{operendJobRunSubtaskList}
\title{Class to contain a list of JobRun subtasks}
\usage{
operendJobRunSubtaskList(...)
}
\arguments{
\item{\dots}{Objects of class \code{\linkS4class{operendJobRunSubtask}}}
}
\description{
This class is a container for one or more
\code{\linkS4class{operendJobRunSubtask}} objects.
}
\section{Extends}{

This class directly extends class \code{\linkS4class{SimpleList}}.
}

\author{
Adam C. Gower \email{agower@bu.edu}
}
