% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{operendPermissions-class}
\alias{operendPermissions-class}
\alias{operendPermissions}
\title{Class to contain permissions for a record}
\usage{
operendPermissions(..., `_other` = character())
}
\arguments{
\item{\dots}{Arguments, named for each group, containing permissions for that group}

\item{_other}{Argument containing permissions for all other groups}
}
\description{
This S4 class is a container for a \code{\linkS4class{SimpleList}} of
character vectors containing permissions associated with an Operend record.
Each element of the list is named for the Group to which the permissions
apply, and may contain:
\describe{
  \item{\code{"C"}}{Create}
  \item{\code{"R"}}{Read}
  \item{\code{"U"}}{Update}
  \item{\code{"D"}}{Delete}
  \item{\code{"PR"}}{Permissions Read}
  \item{\code{"PU"}}{Permissions Update}
}
The list also contains a special element, named \code{`_other`},
that contains the permissions granted to Users in any remaining groups;
this element defaults to an empty vector (i.e., no permissions).
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
