% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operendProcessing.R
\name{Operend processing}
\alias{Operend processing}
\alias{operendPreprocess}
\alias{operendPostprocess}
\title{Pre- or post-process lists during Operend operations}
\usage{
operendPreprocess(x)

operendPostprocess(x)
}
\arguments{
\item{x}{A list object (for \code{operendPreprocess}, must be named)}
}
\value{
\describe{
  \item{\code{operendPreprocess}}{
    The list \code{x}, modified as needed for upload.
  }
  \item{\code{operendPostprocess}}{
    If the list \code{x} has no names, it represents a list of records, and
    a list of the same length as \code{x} will be returned;
    otherwise, a single S4 object will be returned.
  }
}
}
\description{
These functions pre-process lists that will be converted to JSON during PUT
or POST API calls, or post-process list objects converted from JSON in the
HTTP response body of API calls.  These are utility functions that are
called by other functions, and should not be called directly by the
user.
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
