% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{SimpleList to data frame}
\alias{SimpleList to data frame}
\alias{operendSimpleListToDataFrame}
\title{Coerce a SimpleList to a data frame}
\usage{
operendSimpleListToDataFrame(from)
}
\arguments{
\item{from}{A \code{\linkS4class{SimpleList}}}
}
\value{
A data frame with one row for each element in \code{from}
and one column for each slot in \code{from}.
}
\description{
This function coerces a SimpleList object to a data frame.
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
