% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{operendToken-class}
\alias{operendToken-class}
\alias{operendToken}
\title{Class to contain an Operend token}
\description{
This class is an S4 representation of a token generated by an Operend Core
server.
}
\section{Slots}{

\describe{
\item{\code{username}}{A character string specifying the name of the user associated with the token}

\item{\code{name}}{A character string specifying the name of the token}

\item{\code{authorizations}}{A character string specifying the authorizations associated with the token}

\item{\code{creationDate}}{A character string specifying the date the token was created}

\item{\code{secret}}{A character string encoding a JSON object that contains the information
needed to authenticate a call via the Operend API}
}}

\author{
Adam C. Gower \email{agower@bu.edu}
}
