% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{operendUser-class}
\alias{operendUser-class}
\alias{operendUser}
\title{Class to contain data about a User}
\description{
This class is the S4 representation of the User record type.
It is a container for data about a given user.
}
\section{Slots}{

\describe{
\item{\code{username}}{A character string specifying the name of the user.
Defaults to \code{""}.}

\item{\code{firstName}}{A character string specifying the first name of the user.
Defaults to \code{""}.}

\item{\code{lastName}}{A character string specifying the last name of the user.
Defaults to \code{""}.}

\item{\code{email}}{A character string specifying the email address of the user.
Defaults to \code{""}.}

\item{\code{active}}{A logical value specifying whether the user is active.
Defaults to \code{FALSE}.}

\item{\code{superuser}}{A logical value specifying whether the user is a superuser.
Defaults to \code{FALSE}.}

\item{\code{isReadOnlyUser}}{A logical value specifying whether the user has read-only access.
Defaults to \code{FALSE}.}

\item{\code{confirmed}}{A logical value specifying whether the user has been confirmed.
Defaults to \code{FALSE}.}

\item{\code{lastLogin}}{A character string specifying the timestamp of the User's last
authentication. Automatically updated by the system.}

\item{\code{dateJoined}}{A character string specifying the date that the User record was
created. Automatically populated when the record is created.}

\item{\code{group}}{A character string specifying the name of the default Group associated with
the user. Defaults to \code{""}.}

\item{\code{groups}}{A character array specifying the name(s) of the Group(s) associated with
the user. Defaults to an empty character vector.}

\item{\code{canAuthViaPassword}}{A logical value specifying whether the user can authenticate using
a username and password. Defaults to \code{FALSE}.}

\item{\code{canAuthViaGoogle}}{A logical value specifying whether the user can authenticate using
Google credentials. Defaults to \code{FALSE}.}

\item{\code{canAuthViaMicrosoft}}{A logical value specifying whether the user can authenticate using
Microsoft credentials. Defaults to \code{FALSE}.}

\item{\code{canAuthViaCILogon}}{A logical value specifying whether the user can authenticate using
CILogon credentials. Defaults to \code{FALSE}.}

\item{\code{hasMicrosoftOIDCSub}}{A logical value specifying whether the user has a Microsoft OIDC
(OpenID Connect) subscription. Defaults to \code{FALSE}.}

\item{\code{hasCILogonOIDCSub}}{A logical value specifying whether the user has a Microsoft OIDC
(OpenID Connect) subscription. Defaults to \code{FALSE}.}
}}

\seealso{
\describe{
  \item{group}{
    This slot corresponds to the \code{\linkS4class{operendGroup}} object
    slot \code{name}.
  }
}
Functions for working with User records are described in \code{\link{Users}}.
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
