% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{operendUserList-class}
\alias{operendUserList-class}
\alias{operendUserList}
\title{Class to contain a list of User objects}
\usage{
operendUserList(...)
}
\arguments{
\item{\dots}{Objects of class \code{\linkS4class{operendUser}}}
}
\description{
This class is a container for one or more User objects.
}
\section{Extends}{

This class directly extends class \code{\linkS4class{SimpleList}}.
}

\author{
Adam C. Gower \email{agower@bu.edu}
}
