% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{operendVariables-class}
\alias{operendVariables-class}
\alias{operendVariables}
\title{Class to contain Variables records}
\usage{
operendVariables(...)
}
\arguments{
\item{\dots}{Objects of class \code{\linkS4class{operendVariable}}}
}
\description{
This class is the S4 representation of the Variables
record type.
It is a container for the variables and metadata associated with a collection
of variable definitions in an EntityClass record.
}
\section{Slots}{

\describe{
\item{\code{elementType}}{Contains the character string \code{"Variable"}.}

\item{\code{listData}}{A named list of \code{\linkS4class{operendVariable}} objects.}
}}

\section{Extends}{

Directly extends class \code{\linkS4class{SimpleList}}.
}

\seealso{
Variables objects are contained in
\code{\linkS4class{operendEntityClass}} objects.
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
