% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{operendWorkFile-class}
\alias{operendWorkFile-class}
\alias{operendWorkFile}
\title{Class to provide a read-only connection to a WorkFile}
\usage{
operendWorkFile(id, open = c("", "r", "rt", "rb"))
}
\arguments{
\item{id}{An integer-coercible value specifying the unique identifier of a
\code{WorkFile}.}

\item{open}{A character string containing one of the following:
\describe{
  \item{""}{The connection is not opened.}
  \item{"r, rt"}{The connection is operend in read-only text mode.}
  \item{"rb"}{The connection is operend in read-only binary mode.}
}
An \code{operendWorkFile} may not be opened for writing or appending.}
}
\value{
An \code{operendWorkFile} object, containing
a read-only \code{\link[base]{url}} connection to the specified WorkFile,
opened for reading in text or binary mode if specified.
}
\description{
This class is a container for a \code{\link[base]{url}} object
that allows a token-based, read-only connection to a WorkFile.
}
\section{Extends}{

This class directly extends class \code{\link[base]{url}}.
}

\author{
Adam C. Gower \email{agower@bu.edu}
}
