% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{operendWorkFile methods}
\alias{operendWorkFile methods}
\alias{close,operendWorkFile-method}
\alias{gzcon}
\alias{gzcon,operendWorkFile-method}
\alias{open,operendWorkFile-method}
\alias{readBin}
\alias{readBin,operendWorkFile-method}
\alias{readLines}
\alias{readLines,operendWorkFile-method}
\alias{readRDS}
\alias{readRDS,operendWorkFile-method}
\alias{read.table}
\alias{read.table,operendWorkFile-method}
\alias{read.csv}
\alias{read.csv,operendWorkFile-method}
\alias{read.csv2}
\alias{read.csv2,operendWorkFile-method}
\alias{read.delim}
\alias{read.delim,operendWorkFile-method}
\alias{read.delim2}
\alias{read.delim2,operendWorkFile-method}
\alias{scan}
\alias{scan,operendWorkFile-method}
\title{Methods for working with operendWorkFile connections}
\usage{
\S4method{close}{operendWorkFile}(con, ...)

gzcon(con, ...)

\S4method{gzcon}{operendWorkFile}(con, ...)

\S4method{open}{operendWorkFile}(con, ...)

readBin(con, ...)

\S4method{readBin}{operendWorkFile}(con, what, n, ...)

readLines(con, ...)

\S4method{readLines}{operendWorkFile}(con, ...)

readRDS(file, ...)

\S4method{readRDS}{operendWorkFile}(file, ...)

read.table(file, ...)

\S4method{read.table}{operendWorkFile}(file, ...)

read.csv(file, ...)

\S4method{read.csv}{operendWorkFile}(file, ...)

read.csv2(file, ...)

\S4method{read.csv2}{operendWorkFile}(file, ...)

read.delim(file, ...)

\S4method{read.delim}{operendWorkFile}(file, ...)

read.delim2(file, ...)

\S4method{read.delim2}{operendWorkFile}(file, ...)

scan(file, ...)

\S4method{scan}{operendWorkFile}(file, ...)
}
\arguments{
\item{con}{An \code{\linkS4class{operendWorkFile}} object}

\item{\dots}{Arguments to be passed to lower-level functions}

\item{what}{(For \code{readRDS})
Either an object whose mode will give the mode of the vector to be read,
or a character vector of length 1 describing the mode: one of "numeric",
"double", "integer", "int", "logical", "complex", "character", "raw"}

\item{n}{(For \code{readRDS})
A numeric value indicating the maximal number of records to be read;
storage is reserved for \code{n} items.}

\item{file}{An \code{\linkS4class{operendWorkFile}} object}
}
\value{
\describe{
  \item{\code{close}}{
    A NULL value.
  }
  \item{\code{gzcon}}{
    An object inheriting from class \code{"connection"}, with the same
    connection \emph{number} as the input, but with a modified internal
    structure. It has binary mode.
  }
  \item{\code{readBin}}{
    A vector of type \code{what} and length \code{n},
    containing the contents of the WorkFile.
  }
  \item{\code{readLines}}{
    A character vector containing the contents of the lines of the WorkFile.
  }
  \item{\code{readRDS}}{
    The R object encoded in the WorkFile.
  }
  \item{
    \code{read.table},
    \code{read.csv}, \code{read.csv2}, \code{read.delim}, \code{read.delim2}
  }{
    A data frame containing the contents of the WorkFile.
  }
  \item{\code{scan}}{
    A vector containing the contents of the WorkFile.
  }
}
}
\description{
These methods provide functionality for low-level R functions to access a
WorkFile through a read-only connection.
}
\section{Methods}{

Methods are provided here for:
\strong{Connection handlers}
\itemize{
  \item{\code{\link[base]{close}}}
  \item{\code{\link[base]{gzcon}}}
  \item{\code{\link[base]{open}}}
}
\strong{Reading functions}
\itemize{
  \item{\code{\link[base]{readBin}}}
  \item{\code{\link[base]{readLines}}}
  \item{\code{\link[base]{readRDS}}}
  \item{\code{\link[utils]{read.table}}}
  \item{\code{\link[utils]{read.csv}}}
  \item{\code{\link[utils]{read.csv2}}}
  \item{\code{\link[utils]{read.delim}}}
  \item{\code{\link[utils]{read.delim2}}}
  \item{\code{\link[base]{scan}}}
}
}

\author{
Adam C. Gower \email{agower@bu.edu}
}
