% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{operendWorkFileIdList-class}
\alias{operendWorkFileIdList-class}
\alias{operendWorkFileIdList}
\title{Class to contain a list of vectors of WorkFile IDs}
\usage{
operendWorkFileIdList(...)
}
\arguments{
\item{\dots}{Integer vectors}
}
\description{
This class is a container for one or more integer vectors of \code{WorkFile}
identifers.
}
\section{Extends}{

This class directly extends class \code{\linkS4class{SimpleList}}.
}

\author{
Adam C. Gower \email{agower@bu.edu}
}
