% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{operendWorkFileProperties-class}
\alias{operendWorkFileProperties-class}
\alias{operendWorkFileProperties}
\title{Class to contain WorkFile metadata}
\description{
This class is the S4 representation of the WorkFileProperties record
type.  It is a container for metadata about a given WorkFile.
}
\section{Slots}{

\describe{
\item{\code{id}}{An integer value specifying the unique identifier of the WorkFile.
Automatically populated when the file is uploaded.}

\item{\code{hash}}{A character string specifying the MD5 checksum of the file.
Automatically populated when the file is uploaded.}

\item{\code{creator}}{A character string specifying the \code{username} of the User associated with
the record. Automatically populated when the file is uploaded.}

\item{\code{storage}}{A character string specifying the location in which the WorkFile is stored.
This is supplied by the user when the file is uploaded.}

\item{\code{creatorJobRun}}{An integer value specifying the identifier of the JobRun that created this
WorkFile.}

\item{\code{originalName}}{A character string specifying the original name of the file.
Automatically populated when the file is uploaded.}

\item{\code{originalModifiedTime}}{An \code{\linkS4class{operendDate}} object specifying the original timestamp
of the file.}

\item{\code{fileType}}{A character string specifying the type (e.g., extension) of the file.
Created when the file is uploaded.}

\item{\code{isTrashed}}{A logical value specifying whether the WorkFile is flagged to be
trashed. Defaults to \code{FALSE}.}

\item{\code{isTransient}}{A logical value specifying whether the WorkFile is flagged as
temporary. Defaults to \code{FALSE}.}

\item{\code{creationDatetime}}{A character string specifying the date and time that the record was created.
Automatically populated when the file is uploaded.}

\item{\code{length}}{A numeric value specifying the length of the file in bytes.
Automatically populated when the file is uploaded.}

\item{\code{token}}{A character string specifying a token that can be used for password-free
authentication. Automatically populated when the file is uploaded.}

\item{\code{permissions}}{An \code{\linkS4class{operendPermissions}} object specifying the permissions
associated with the record. Defaults to read-only.}
}}

\seealso{
Functions for working with WorkFileProperties records are described in
\code{\link{WorkFileProperties}}.
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
