% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{operendWorkFilePropertiesList-class}
\alias{operendWorkFilePropertiesList-class}
\alias{operendWorkFilePropertiesList}
\title{Class to contain a list of WorkFileProperties objects}
\usage{
operendWorkFilePropertiesList(...)
}
\arguments{
\item{\dots}{Objects of or extending class \code{\linkS4class{operendWorkFileProperties}}}
}
\description{
This class is a container for one or more WorkFileProperties objects.
}
\section{Extends}{

This class directly extends class \code{\linkS4class{SimpleList}}.
}

\author{
Adam C. Gower \email{agower@bu.edu}
}
