% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{opeRend S4 show methods}
\alias{opeRend S4 show methods}
\alias{show,operendDate-method}
\alias{show,operendEntity-method}
\alias{show,operendEntityList-method}
\alias{show,operendEntityClass-method}
\alias{show,operendEntityClassList-method}
\alias{show,operendGroup-method}
\alias{show,operendJobRun-method}
\alias{show,operendPermissions-method}
\alias{show,operendToken-method}
\alias{show,operendUser-method}
\alias{show,operendWorkFileProperties-method}
\title{Methods for showing opeRend S4 objects}
\usage{
\S4method{show}{operendDate}(object)

\S4method{show}{operendEntity}(object)

\S4method{show}{operendEntityList}(object)

\S4method{show}{operendEntityClass}(object)

\S4method{show}{operendEntityClassList}(object)

\S4method{show}{operendGroup}(object)

\S4method{show}{operendJobRun}(object)

\S4method{show}{operendPermissions}(object)

\S4method{show}{operendToken}(object)

\S4method{show}{operendUser}(object)

\S4method{show}{operendWorkFileProperties}(object)
}
\arguments{
\item{object}{An R object}
}
\value{
\code{show} returns an invisible \code{NULL}.
}
\description{
These methods show the contents of various S4 objects in a human-readable
format.
}
\section{Methods}{

Class-specific methods exist for:
\itemize{
  \item \code{\linkS4class{operendDate}}
  \item \code{\linkS4class{operendEntity}}
  \item \code{\linkS4class{operendEntityList}}
  \item \code{\linkS4class{operendEntityClass}}
  \item \code{\linkS4class{operendEntityClassList}}
  \item \code{\linkS4class{operendGroup}}
  \item \code{\linkS4class{operendJobRun}}
  \item \code{\linkS4class{operendPermissions}}
  \item \code{\linkS4class{operendToken}}
  \item \code{\linkS4class{operendUser}}
  \item \code{\linkS4class{operendWorkFileProperties}}
}
}

\author{
Adam C. Gower \email{agower@bu.edu}
}
