% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Users.R
\name{Users}
\alias{Users}
\alias{addUser}
\alias{getUser}
\alias{listUsers}
\alias{updateUser}
\title{Add, retrieve, update, or list User record(s)}
\usage{
addUser(username, ..., verbosity = getOption("opeRend")$verbosity)

getUser(username)

listUsers(active = TRUE, asDataFrame = TRUE)

updateUser(username, ..., verbosity = getOption("opeRend")$verbosity)
}
\arguments{
\item{username}{A character string specifying the name of a \code{User}
record to be added, retrieved, or updated.}

\item{\dots}{Additional arguments specifying fields of the \code{User}
record to be added or updated, or fields on which to limit a listing}

\item{verbosity}{A value coercible to a nonnegative integer, specifying the verbosity level.
A value of FALSE, TRUE, 0 or 1 does not produce any messages.
A value of 2 instructs the curl calls to produce verbose output.
A value greater than 2 produces additional output.
Defaults to \code{getOption("opeRend")$verbosity}.}

\item{active}{A logical value specifying whether to list only \code{User}s that are
active or inactive; a value of \code{NA} shows all \code{User}s
regardless of active status.}

\item{asDataFrame}{A logical value specifying whether to return the listing as a data frame.
Defaults to \code{TRUE}.}
}
\value{
\describe{
  \item{\code{addUser}, \code{getUser}, \code{updateUser}}{
    If the operation is successful,
    an \code{\linkS4class{operendUser}} object.
  }
  \item{\code{listUsers}}{
    \describe{
      \item{If \code{asDataFrame} = \code{TRUE}}{
        A data frame containing one row per record and one column per field.
      }
      \item{If \code{asDataFrame} = \code{FALSE}}{
        An \code{\linkS4class{operendUserList}} object.
      }
    }
  }
  \item{All functions}{
    If an error is encountered, the function terminates with a message.
  }
}
}
\description{
These functions attempt to add or retrieve a \code{User} record,
or to list \code{User} records.
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
