% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorkFiles.R
\name{WorkFiles}
\alias{WorkFiles}
\alias{addWorkFile}
\alias{getWorkFile}
\alias{purgeWorkFile}
\title{Functions for working with WorkFiles}
\usage{
addWorkFile(
  file,
  storage = getOption("opeRend")$storage_location,
  fileType,
  token,
  originalName,
  originalModifiedTime,
  permissions,
  verbosity = getOption("opeRend")$verbosity
)

getWorkFile(id, filename, verbosity = getOption("opeRend")$verbosity)

purgeWorkFile(id, verbosity = getOption("opeRend")$verbosity)
}
\arguments{
\item{file}{One of the following:
\itemize{
  \item{
    a character string containing the full path to a file
  }
  \item{
    a binary-mode connection (which will be opened if necessary,
    and closed at the end of the function call)
  }
}}

\item{storage}{A character string specifying the location in which the WorkFile is stored.}

\item{fileType}{A character string specifying the file type (e.g., extension) of the
\code{WorkFile}}

\item{token}{A character string specifying a token that can be used for password-free
authentication. Automatically populated when a \code{WorkFile} is uploaded.}

\item{originalName}{A character string specifying the original name of the file.  If \code{file}
is a character string, this defaults to the base name of the file.}

\item{originalModifiedTime}{An numeric-coercible value, specifying the original modification time of the
file (in seconds since the Unix epoch).  If \code{file} is a character
string, this defaults to the 'mtime' of the file.}

\item{permissions}{An optional \code{\linkS4class{operendPermissions}} object
specifying the permissions to be used when creating or updating the record}

\item{verbosity}{An integer-coercible value specifying the verbosity level:
\describe{
  \item{\code{0}}{Do not print messages}
  \item{\code{1}}{Print only high-level messages}
  \item{\code{2}}{Show headers}
  \item{\code{3}}{Show headers and bodies}
  \item{\code{4+}}{Show headers, bodies, and curl status messages}
}
Defaults to \code{getOption("opeRend")$verbosity}.}

\item{id}{An integer-coercible value specifying the unique identifier of a
\code{WorkFileProperties} record.
Automatically populated when a \code{WorkFile} is uploaded.}

\item{filename}{A character string specifying the full path to a file}
}
\value{
\describe{
  \item{\code{addWorkFile}}{
    The file is uploaded, and if the operation is successful, a
    \code{linkS4class{operendWorkFileProperties}} object is returned.
  }
  \item{\code{getWorkFile}}{
    No value is returned.
  }
  \item{\code{purgeWorkFile}}{
    A logical value stating whether the operation was successful (invisibly).
  }
  \item{All functions}{
    If an error is encountered, the function terminates with a message.
  }
}
}
\description{
These functions attempt to add or retrieve a \code{WorkFile}
or to purge a \code{WorkFile} from the system.
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
