% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EntityClasses.R
\name{EntityClasses}
\alias{EntityClasses}
\alias{addEntityClass}
\alias{deleteEntityClass}
\alias{getEntityClass}
\alias{listEntityClasses}
\alias{updateEntityClass}
\title{Add, delete, retrieve, update, or list EntityClass record(s)}
\usage{
addEntityClass(file, verbosity = getOption("opeRend")$verbosity)

deleteEntityClass(name, verbosity = getOption("opeRend")$verbosity)

getEntityClass(name, verbosity = getOption("opeRend")$verbosity)

listEntityClasses()

updateEntityClass(file, verbosity = getOption("opeRend")$verbosity)
}
\arguments{
\item{file}{One of the following:
\itemize{
  \item{
    a character string containing the full path to a file
    containing an \code{EntityClass} definition in JSON
  }
  \item{
    a \code{\link{connection}} (which will be opened if necessary, and closed
    at the end of the function call) to a file
    containing an \code{EntityClass} definition in JSON
  }
}}

\item{verbosity}{A value coercible to a nonnegative integer, specifying the verbosity level.
A value of FALSE, TRUE, 0 or 1 does not produce any messages.
A value of 2 instructs the curl calls to produce verbose output.
A value greater than 2 produces additional output, including the value of the
\code{content} parameter.
Defaults to \code{getOption("opeRend")$verbosity}.}

\item{name}{A character string specifying the name of an \code{EntityClass}
record to be deleted or retrieved}
}
\value{
\describe{
  \item{
    \code{addEntityClass}, \code{getEntityClass}, \code{updateEntityClass}
  }{
    If the operation is successful,
    an \code{\linkS4class{operendEntityClass}} object.
  }
  \item{\code{deleteEntityClass}}{
    A logical value stating whether the operation was successful (invisibly).
  }
  \item{\code{listEntityClasses}}{
    An \code{\linkS4class{operendEntityClassList}} object.
  }
  \item{All functions}{
    If an error is encountered, the function terminates with a message.
  }
}
}
\description{
These functions attempt to add, delete, retrieve or update an
\code{EntityClass} record, or to list \code{EntityClass} records.
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
