% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operendApiCall.R
\name{operendApiCall}
\alias{operendApiCall}
\title{Perform an Operend API call}
\usage{
operendApiCall(
  url,
  method = c("GET", "POST", "PUT", "DELETE"),
  accept = c("application/json", "application/octet-stream"),
  content,
  contentType = c("application/json", "application/octet-stream"),
  verbosity = getOption("opeRend")$verbosity
)
}
\arguments{
\item{url}{A character string specifying the URL to which an API call will be made}

\item{method}{A character string specifying the method to be used to submit the API call}

\item{accept}{A character string specifying the "Accept" HTTP header to be passed to
\code{curlPerform}; defaults to \code{"application/json"}}

\item{content}{An optional variable specifying the content to be uploaded during a POST or
PUT request}

\item{contentType}{A character string specifying the "Content-Type" HTTP header to be passed to
\code{curlPerform}; defaults to \code{"application/json"}}

\item{verbosity}{A logical or nonnegative numeric value specifying the verbosity level.
A value of FALSE, TRUE, 0 or 1 does not produce any messages.
A value of 2 instructs the curl calls to produce verbose output.
A value greater than 2 produces additional output, including the value of the
\code{content} parameter.
Defaults to \code{getOption("opeRend")$verbosity}.}
}
\value{
If an error is encountered, an error is thrown,
containing a message constructed from the HTTP response body.
Otherwise:
\itemize{
  \item{
    If the \code{Content-Type} attribute of the HTTP response body is set to
    \code{'application/json'}, the response body is converted from JSON to an
    R object, and that object is returned.
  }
  \item{
    If that attribute is set to anything else (or does not exist), the HTTP
    response body itself is returned.
  }
}
}
\description{
This function uses \code{\link{curlPerform}} to carry out Operend API calls,
converting all input and output to/from JSON as needed.
}
\details{
\code{\link{curlPerform}} is used in place of \code{\link{getURL}}
(which does not collect the HTTP header) and \code{\link{getURLContent}}
(which does not return the HTTP response body in the event of an HTTP error)
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
